refactored update_project_etag and callbacks into shared concern ProjectUpdater
