+++ app/models/project_permission.rb
@@ -1,9 +1,8 @@
 class ProjectPermission < ActiveRecord::Base
   include RequestAudited
+  include ProjectUpdater
   default_scope { order('created_at DESC') }
   audited
-  after_save :update_project_etag, if: :saved_changes?
-  after_destroy :update_project_etag
 
   belongs_to :user
   belongs_to :project
@@ -13,14 +12,4 @@ class ProjectPermission < ActiveRecord::Base
   validates :user_id, presence: true, uniqueness: {scope: :project_id, case_sensitive: false}
   validates :project_id, presence: true
   validates :auth_role_id, presence: true
-
-  private
-
-  def update_project_etag
-    last_audit = self.audits.last
-    new_comment = last_audit.comment ? last_audit.comment.merge({raised_by_audit: last_audit.id}) : {raised_by_audit: last_audit.id}
-    self.project.update(etag: SecureRandom.hex, audit_comment: new_comment)
-    last_parent_audit = self.project.audits.last
-    last_parent_audit.update(request_uuid: last_audit.request_uuid)
-  end
 end
